add test for NewRedisClient
